refactor: Split out choice and sequence parsers
